add ipaaca (throws connection error), modify selection of sentence (chooses sentence wasting least information) and weighting @generate.py, integrate modal / direction information @application.py. adjust to python 2.

+++ parse.py
@@ -44,7 +44,7 @@ def parse_string(base_dir, string):
             indexes.append(0)
         elif ind < last_ind:  # Up to next item in parent list
             diff = (last_ind - ind)
-            for i in range(last_ind - ind):
+            for i in range((int)(last_ind - ind)):
                 level -= 1
                 indexes.pop()
             indexes[level] += 1
